refactor into postDrafts.ts
